Added a simple built-in editor.
The editor widget is a modified version of Ted Conbeer's widget.

    https://github.com/tconbeer/textual-textarea

I have made quite a few changes to provide a simpler editor, prevent
complaints from linting tools and, in some places, bring things in line
with my coding style.

This widget is not intended to be used long term. I hope that the
planned Textual TextArea widget will replace it in due course.